+++ lib/render/FlowRenderer.cpp
@@ -295,7 +295,7 @@ FlowRenderer::_paintGL( bool fast )
             if( params->GetFlowDirection() == 1 )           // backward integration
                 deltaT *= -1.0f;
             int numOfSteps = params->GetSteadyNumOfSteps();
-            for( int i = _advection.GetMaxNumOfPart() - 1;  // existing number of advection steps 
+            for( size_t i = _advection.GetMaxNumOfPart() - 1;  // existing number of advection steps 
                  i < numOfSteps && rv == flow::ADVECT_HAPPENED; i++ )
             {
                 rv = _advection.AdvectOneStep( &_velocityField, deltaT );
@@ -307,7 +307,7 @@ FlowRenderer::_paintGL( bool fast )
                 assert( deltaT > 0.0f );
                 float   deltaT2 = deltaT * -1.0f;
                 rv = flow::ADVECT_HAPPENED;
-                for( int i = _2ndAdvection->GetMaxNumOfPart() - 1; 
+                for( size_t i = _2ndAdvection->GetMaxNumOfPart() - 1; 
                      i < numOfSteps && rv == flow::ADVECT_HAPPENED; i++ )
                 {
                     rv = _2ndAdvection->AdvectOneStep( &_velocityField, deltaT2 );
